Changed Pack to Sellable & UML diagram

+++ app/main/service/pack_service.py
@@ -1,5 +1,5 @@
 from app.main import db
-from app.main.model.pack import Pack
+from app.main.model.sellable import Pack
 
 
 def get_packs():
@@ -13,9 +13,17 @@ def get_pack(pack_id):
 def save_new_pack(data):
     new_pack = Pack(
         name=data['name'],
-        description=data['description']
+        description=data['description'],
+        image_str=data['image_str'],
+        stock=data['stock'],
+        price=data['price'],
+        category=data['category'],
+        year=data['year'],
+        team=data['team'],
+        type=data['type']
     )
-    save_changes(new_pack)
+    db.session.add(data)
+    db.session.commit()
     response_object = {
         'id': new_pack.id,
     }
@@ -25,26 +33,44 @@ def save_new_pack(data):
 def update_packs(data):
     query = db.session.query(Pack)
     for pack in data:
-        new_query = query.filter(Pack.id == pack['id'])
-        record = new_query.one()
-        record.name = pack["name"]
-        record.description = pack["description"]
+        record = query.filter(Pack.id == pack['id']).first()
+
+        if not record:
+            response_object = {
+                'status': 'fail',
+                'message': 'Pack with id ' + str(pack['id']) + ' does not exist.'
+            }
+            return response_object, 409
+
+        record.name = pack['name'],
+        record.description = pack['description'],
+        record.image_str = pack['image_str'],
+        record.stock = pack['stock'],
+        record.price = pack['price'],
+        record.category = pack['category'],
+        record.year = pack['year'],
+        record.team = pack['team'],
+        record.type = pack['type']
+
     db.session.flush()
     db.session.commit()
     response_object = {
         'status': 'success',
-        'message': 'Successfully updated row.',
+        'message': 'Successfully updated rows.',
     }
     return response_object, 201
 
 
-def save_changes(data):
-    db.session.add(data)
-    db.session.commit()
-
-
 def delete_pack(data):
-    pack = Pack.query.filter_by(id=data['id']).one()
+    pack = Pack.query.filter_by(id=data['id']).first()
+
+    if not pack:
+        response_object = {
+            'status': 'fail',
+            'message': 'Pack with id ' + str(data['id']) + ' does not exist.'
+        }
+        return response_object, 409
+
     db.session.delete(pack)
     db.session.commit()
-    return {'message': "Pack id : " + str(data['id']) + " was deleted successfully"}, 200
+    return {'message': "Pack with id " + str(data['id']) + " was deleted successfully."}, 200
